Mapeo y llamado general al SOAP
Realice el mapeo y la llamada al soap del ahi va. Segun estuve viendo solo debo terminar de definir las credenciales de forma correcta

+++ AirTableApi/Dtos/Ahiva/DtoPaquete.cs
@@ -2,7 +2,7 @@
 {
     public class DtoPaquete
     {
-        public decimal Peso { get; set; }
+        public double Peso { get; set; }
         public string ResponsableServEntrega { get; set; }
         public string Empaque { get; set; }
         public string Almacenamiento { get; set; }
